internal/postgres: update symbol_history when version is the same
shouldUpdateSymbolHistory now returns true when the old and new versions
are the same.

That way if we make a change to a column for that module version, such
as most recently the package_symbol_id, we can just refetch the module
to update the symbol_history row.

For golang/go#37102

Change-Id: I1faca7aa09b1f244f99ffd33679231e1104793ca
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/315311
Trust: Julie Qiu <julie@golang.org>
Reviewed-by: Jonathan Amsterdam <jba@google.com>