internal/fronted: update symbol link for multi GOOS
When a symbol is introduced for a specific GOOS/GOARCH at a version,
linking to an unspecified GOOS/GOARCH page might not take the user to
the symbol. Instead, link to one of the supported build contexts.

For golang/go#37102

Change-Id: I2b4edc6c7bd12e148d6560b0e84c005f6511a0d3
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/315429
Trust: Julie Qiu <julie@golang.org>
Run-TryBot: Julie Qiu <julie@golang.org>
Reviewed-by: Jonathan Amsterdam <jba@google.com>