internal/postgres: mark processed modules from index with status 0
When a module is added for the first time using InsertIndexVersions,
set the status to 0, even if the module already exists in
module_version_states. That way we will prioritize reprocessing this
module, in case its previous status was a non-2xx code.

For golang/go#46117

Change-Id: Ie2c01a84449a0bb2e446a0c1388b315b580d3b12
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/319049
Trust: Julie Qiu <julie@golang.org>
Reviewed-by: Jonathan Amsterdam <jba@google.com>