content,internal: create styleguide page
The content of the styleguide page is generated from
the markdown files contained in the component directories
of content/static. We use goldmark to parse the files
into html sections and generate an outline for the sidenav.
Goldmark is extended to render html code blocks as both
html and code snippets.

Change-Id: I023edb77114001439be58e00d5a48198181f165a
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/320654
Trust: Jamal Carvalho <jamal@golang.org>
Run-TryBot: Jamal Carvalho <jamal@golang.org>
Reviewed-by: Julie Qiu <julie@golang.org>