internal/fetch: skip large module forks
If the worker sees a module without a go.mod file that is identical to
a one of a hardcoded list of known modules but has a different module
path, then the worker will mark it as an alternative module and avoid
processing it.

This implements the scheme described in the issue, except (1) we use
the AlternativeModule status code for this case instead of a new code,
and (2) we require that the versions match in addition to the zip
contents.

This first CL has the signatures for aws-sdk-go. We will add
more in subsequent CLs.

For golang/go#46341

Change-Id: I6d429464fcb48da43f9c874bc9b37049466bafb6
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/322529
Trust: Jonathan Amsterdam <jba@google.com>
Run-TryBot: Jonathan Amsterdam <jba@google.com>
TryBot-Result: kokoro <noreply+kokoro@google.com>
Reviewed-by: Julie Qiu <julie@golang.org>