migrations: enlarge package_symbols references to paths
The package_symbols table has two columns that reference paths by id.
Enlarge them from integer to bigint.

Since package_symbols is unused in prod, we do this directly. That
will lock the table for about an hour (based on running this migration
in exp).

I discovered that (at least on the test DBs) it was necessary to drop
and re-create one of the indexes on the columns (but strangely, not
the others).

Change-Id: I9698c8e4a5ce1670469fc3dcc6ebac6b6c4f4d7d
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/323850
Trust: Jonathan Amsterdam <jba@google.com>
Run-TryBot: Jonathan Amsterdam <jba@google.com>
TryBot-Result: kokoro <noreply+kokoro@google.com>
Reviewed-by: Jamal Carvalho <jamal@golang.org>