internal/database:  reduce max retries on serialization failure
If a save-module transaction fails due to serialization, then most
likely another version of the same module is being processed at the
same time.

The right thing to do is to retry the transaction, and we do, up to 30
times. But if many versions are being reprocessed at once--as seems to
be frequently the case now--we are just wasting time and compute
resources, because we retry immediately, without any backoff. It would
be better to fail sooner, and let the natural exponential backoff of
the task queue do the retrying for us.

We don't want to eliminate the immediate retry completely, but reduce
the count to 10.

Change-Id: I89f1cb8a9c021083f838c6659d3e5b39226cb144
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/323411
Trust: Jonathan Amsterdam <jba@google.com>
Run-TryBot: Jonathan Amsterdam <jba@google.com>
TryBot-Result: kokoro <noreply+kokoro@google.com>
Reviewed-by: Jamal Carvalho <jamal@golang.org>