Revert "internal/postgres: process new modules in arrival order"
This reverts commit 61bac80a3d712e004b0fd1355b057c8905d43dce.

Reason for revert: if we queue in arrival order, then multiple versions of the same module that are published together (which seems to happen a lot) will be queued together, and will conflict with each other on the module path lock.

We'll have to remove the processing lag alert, but instead we'll alert on total backlog, which is more robust anyway.

Change-Id: Ia1f9f5c7ec8b0b6577527d3dd6060b692c7a7b28
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/323172
Trust: Jonathan Amsterdam <jba@google.com>
Reviewed-by: Julie Qiu <julie@golang.org>