devtools/ci: add files for CI
Files used for kokoro are copied to devtools/ci along with a README.md,
so that it is clear what these files are used for.

The existing files will be deleted in the next CL, after an internal CL
is merged to change references.

Change-Id: I893ac9461d3f775460ba749da80e2b4ba71dcc1e
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/325129
Trust: Julie Qiu <julie@golang.org>
Reviewed-by: Jonathan Amsterdam <jba@google.com>