devtools,internal: build new static assets
The static build devtool builds the TypeScript files
in static and ignores content/static. Unused config
options like Write are removed and support is added
for bundling files together. The command walks a
given directory and collects files it determines to
be entry points. Files prefixed with an underscore
are assumed to be imported in an entrypoint file
and ignored.

Change-Id: I7c62afd36ff4ee327d528919c54866631a9c73f7
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/327281
Trust: Jamal Carvalho <jamal@golang.org>
Reviewed-by: Julie Qiu <julie@golang.org>