internal/{postgres,frontend}: fix search grouping
Grouping of search results was broken (the sort predicate was not
transitive). This CL fixes it.

Raises the score of the first package in the module with the highest
major version of its series so that it appears first in the list of
search results.

Also, replace the list of search results of lower-major-version
modules with a set of module paths at _all_ other major versions.

Make corresponding changes to the frontend code and templates.

Change-Id: I2236baa00d4edacf65246e2f65809a1f74b231c1
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/328335
Trust: Jonathan Amsterdam <jba@google.com>
Run-TryBot: Jonathan Amsterdam <jba@google.com>
TryBot-Result: kokoro <noreply+kokoro@google.com>
Reviewed-by: Julie Qiu <julie@golang.org>