internal/postgres: carefully change status on index insert
We've been updating the status to 0 on every module that comes from
the index, even if it's already in the DB with a different status.

While we do want to change the status when it's reasonable to
reprocess (see the linked issue below), we don't in most other cases.

So only update the status when we see a 404.

For golang/go#46117

Change-Id: I2774a0e5637873754cf69805bf154aaab632b0b8
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/328929
Trust: Jonathan Amsterdam <jba@google.com>
Run-TryBot: Jonathan Amsterdam <jba@google.com>
Reviewed-by: Julie Qiu <julie@golang.org>
TryBot-Result: kokoro <noreply+kokoro@google.com>