devtools/docker: run chrome with nodejs container
The nodejs service now depends on headless chrome to be up an running,
so that it can be used to run e2e tests independent of the frontend
service.

This is used to run npx in e2e tests using a different instance of the
one running in docker, such as when testing against a local instance or
when running the deployment script.

Change-Id: I68d583859a7abe8cf7b59d5a79eebe3d994c54d5
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/329915
Trust: Julie Qiu <julie@golang.org>
Run-TryBot: Julie Qiu <julie@golang.org>
Reviewed-by: Jonathan Amsterdam <jba@google.com>