internal/postgres: change symbolsearch to index underscores
The symbol search insert and update queries are changed to index
underscores as slashes. For example, "A_B" is indexed as "A/B".

Underscores are treated as "blanks" by the postgres parser. However, we
want "A_B" to be ranked lower than just "A" in a search for "A", not
equally.

For golang/go#44142

Change-Id: I4399c046adae33f69cb806012a68127ab21cb937
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/332413
Trust: Julie Qiu <julie@golang.org>
Run-TryBot: Julie Qiu <julie@golang.org>
TryBot-Result: kokoro <noreply+kokoro@google.com>
Reviewed-by: Jonathan Amsterdam <jba@google.com>