internal/postgres/{symbolsearch}: add QuerySymbol
A new internal/postgres/symbolsearch package is added, which is used to
construct queries for symbol search.

Queries are generated in TestGenerateQuery, and writted to query.go.
These are then used in internal/postgres. The query to be used is chosen
based on the search input.

In this CL, only QuerySymbol is added, which will search for a symbol
based on an exact match on the symbol name.

For golang/go#44142

Change-Id: I0c4514f6cfc0b053faf732d7367418d096771145
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/335262
Trust: Julie Qiu <julie@golang.org>
Run-TryBot: Julie Qiu <julie@golang.org>
TryBot-Result: kokoro <noreply+kokoro@google.com>
Reviewed-by: Jonathan Amsterdam <jba@google.com>