internal/postgres: change experiment for documentation_symbols upsert
Data is now inserted into documentation_symbols if the
insert-symbol-search-documents experiment is on, as opposed to if
skip-insert-symbols is off.

This table is needed for the symbol search design, but it's unnecessary
to insert into that table unless we are also inserting into symbol
search documents.

The experiment is added to various symbol history tests, since it uses
data from that table as a check against data in the symbol_history
table.

For golang/go#44142

Change-Id: I24f1689a9456e555274bf1c6725e917a8e9a8d45
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/335266
Trust: Julie Qiu <julie@golang.org>
Run-TryBot: Julie Qiu <julie@golang.org>
TryBot-Result: kokoro <noreply+kokoro@google.com>
Reviewed-by: Jamal Carvalho <jamal@golang.org>