tests/e2e: fix errors test
The errors package test is updated to account for the new
behavior of the left nav in which it automatically selects
and expands sections as they are scrolled into view. Instead
of clicking the documentation section link the tests waits
for that section to expand.

Change-Id: Ic37afde43622799cb311cdbd4105408d2126e8f4
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/336232
Trust: Jamal Carvalho <jamal@golang.org>
Reviewed-by: Julie Qiu <julie@golang.org>
Run-TryBot: Jamal Carvalho <jamal@golang.org>
TryBot-Result: kokoro <noreply+kokoro@google.com>