internal/postgres: handle subpaths in multiword searches
When a word in a multiword search contains a known hostname or a "/", we
know that that word cannot be a symbol name. Instead, ensure that the
tsv_path_tokens search matches on that element.

For golang/go#44142

Change-Id: I50a29bb647ad96f4f97ec06f206993619e1f0fb0
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/338170
Trust: Julie Qiu <julie@golang.org>
Run-TryBot: Julie Qiu <julie@golang.org>
TryBot-Result: kokoro <noreply+kokoro@google.com>
Reviewed-by: Jonathan Amsterdam <jba@google.com>