internal/postgres/symbolsearch: handle multiword search better
The logic for multiple words is changed such that words that do not
match the symbol name much match on path tokens.

For example, if a user searches for "julieqiu foo", previously the
symbol "foo" in package "foo" would be returned. Now, if the symbol name
is "foo", tsv_path_tokens @@ <query> must return true for "julieqiu" in
order for a result to be returned.

For a given search, all of the possible search combinations are
generated. If there are more than 2 combinations, an OR search is
performed as was done prior to this CL. Otherwise, each search
combination is performed and results are combined together and sorted by
imported by count.

For golang/go#44142

Change-Id: I3ab45512d19c911c1cf158edade306a4ce67fde6
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/338172
Trust: Julie Qiu <julie@golang.org>
Run-TryBot: Julie Qiu <julie@golang.org>
TryBot-Result: kokoro <noreply+kokoro@google.com>
Reviewed-by: Jamal Carvalho <jamal@golang.org>