migrations: change search_documents primary_key to package_path_id
The PK for search_documents is changed from package_path to
package_path_id, so that package_path_id has a unique constraint that
can be used for a FK reference by
symbol_search_documents.package_path_id. This also has the added benefit
that the PK index should be smaller.

Tests will fail on this CL because there is no longer a UNIQUE
constraint on package_path, but this is fixed in the next CL. The
following CL should be merged at the same time.

For golang/go#44142

Change-Id: If5900db89bb0ee85532525c9fd031136c66640fb
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/339109
Trust: Julie Qiu <julie@golang.org>
Run-TryBot: Julie Qiu <julie@golang.org>
Reviewed-by: Jamal Carvalho <jamal@golang.org>