devtools/cmd/seeddb: process same module sequentially
Process multiple versions of the same module sequentially. Doing so in
parallel causes DB contention because of the lock on the module path
and other ordinary DB locks. This can cause a fetch to reach the limit
on the number of transaction retries.

Change-Id: Ibb58e583441e0b471a628edd1b64ec7f8010e174
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/340118
Trust: Jonathan Amsterdam <jba@google.com>
Run-TryBot: Jonathan Amsterdam <jba@google.com>
Reviewed-by: Julie Qiu <julie@golang.org>