tests/e2e: update jest environment settings
Usage of GO_DISCOVERY_E2E_ENVIRONMENT is updated so that if it is set to
"staging" or "prod", then the staging snapshots are used. Otherwise, the
ci snapshots are used.

A license header is add to tests/e2e/snapshotResolver.js and comments
are added for future readers.

Change-Id: Ib9d0ba2032b27479b3ecab93347fdfb939bd2aee
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/341109
Trust: Julie Qiu <julie@golang.org>
Reviewed-by: Jamal Carvalho <jamal@golang.org>