internal/frontend,static: use singular nouns for search modes
The search modes are changed to the singular "package" and "identifier".
The constant values are also added to basePage, to ensure that the same
string is used throughout the templates and Go code.

For golang/go#47320
For golang/go#44142

Change-Id: Ia82f81b64fb267f4fe352359c0276426a69ebdaa
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/341178
Trust: Julie Qiu <julie@golang.org>
Run-TryBot: Julie Qiu <julie@golang.org>
Reviewed-by: Jamal Carvalho <jamal@golang.org>