internal/postgres/symbolsearch: remove support for matching on field and method names
Symbol search on an exact symbol names that matches on field or method
names is too slow, so support is removed until performance can be
improved.

The cause is because too many symbol ids will be returned (for example,
3 rows on a exact match for "db", vs 5.2K+ when also matching on field
and method names).

For golang/go#44142

Change-Id: Ie59b376a84ea0344a4c40c804b28324d32594e72
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/341670
Trust: Julie Qiu <julie@golang.org>
Reviewed-by: Jonathan Amsterdam <jba@google.com>