internal,static/frontend: change identifier to symbol
All references to identifier are changed to symbol.

A capitalize method is added to the frontend template
functions, which runs strings.Title. This is used to
capitalize the string "symbol" in templates so that we
can use basePage.SearchModeSymbol, instead of a hardcoded
string, in case this changes again in the future.

For golang/go#44142

Change-Id: Ia72c7bd31213c508050cd8756b5eae099736d3ba
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/341869
Reviewed-by: Jonathan Amsterdam <jba@google.com>
Trust: Julie Qiu <julie@golang.org>