internal/frontend: add symbol filters
The # symbol filter is updated so that it can be used with any word,
instead of only as a prefix to indicate the user wants to search in
symbol mode.

For example, "#foo bar" was understood to mean, search for the query
string 'foo bar', in symbol search mode.

This is now understood to mean, search for the query string 'foo bar',
in symbol search mode, where 'foo' MUST match a symbol name.

This syntax allows users to specify which symbol to search for.

postgres.Search will support filters in the next CL.

For golang/go#44142

Change-Id: Ifb69f0d2745d68b5e0c28d734b48e58cc02c1819
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/347610
Trust: Julie Qiu <julie@golang.org>
Run-TryBot: Julie Qiu <julie@golang.org>
Reviewed-by: Jonathan Amsterdam <jba@google.com>
Reviewed-by: Jamal Carvalho <jamal@golang.org>
TryBot-Result: kokoro <noreply+kokoro@google.com>