internal/godoc: remove IsDeprecated from go/doc
pkgsite has a fork of the standard library's go/doc package. This CL
is part of an attempt to reconcile the two so pkgsite can use the
standard library go/doc.

Remove the IsDeprecated field from our go/doc. We can determine
if a symbol is deprecated using its doc, which is exported, so
we don't need to have that logic in the go/doc package itself.

Also, the experiment that displays deprecated symbols is fully
active, so remove code that depends on it.

We will remove the actual experiment definition in a later CL.

Change-Id: I723ffc4ee48c32e791eab896df61cfc389874f98
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/372754
Trust: Jonathan Amsterdam <jba@google.com>
Run-TryBot: Jonathan Amsterdam <jba@google.com>
TryBot-Result: kokoro <noreply+kokoro@google.com>
Reviewed-by: Jamal Carvalho <jamal@golang.org>