internal/fetch: list known alternative module paths
Some modules don't have go.mod files. Nonetheless there is clearly a
canonical module path, and various forks that should be marked as
alternative. A major example is github.com/Azure/azure-sdk-for-go,
the canonical path, and forks like gopkg.in/Azure/azure-sdk-for-go.
For large modules like the Azure SDK, these forks consume a lot of
processing bandwidth.

Although we already have a mechanism for this (see largefork.go), it
requires both contents and version numbers to match a known list of
modules and versions.  The list can get out of date, and it doesn't
include every possible version. For instance, the gopkg.in Azure fork
consists of many pseudo-versions that aren't in the proxy list.

This CL adds a list of module paths that are known to be forks.
They will be marked as alternative modules regardless of version
or contents.

Fixes golang/go#52329.

Change-Id: Ie374da67f0f42f50be551a1078b667686453b447
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/400357
Run-TryBot: Jonathan Amsterdam <jba@google.com>
TryBot-Result: kokoro <noreply+kokoro@google.com>
Reviewed-by: Julie Qiu <julieqiu@google.com>