internal/godoc/dochtml: only recognize "Deprecated" at paragraph start
Treat a function or method as deprecated only if the string
"Deprecated:" occurs at the start of a paragraph.  Previously, we
looked for it at the start of a line.

Also, remove the "deprecated-off" golden file because it was no longer
used.

Fixes golang/go#46913.

Change-Id: I0d16598915677e17565e9944ce67092dcdd208f5
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/400994
Run-TryBot: Jonathan Amsterdam <jba@google.com>
TryBot-Result: kokoro <noreply+kokoro@google.com>
Reviewed-by: Jamal Carvalho <jamal@golang.org>
Reviewed-by: Robert Findley <rfindley@google.com>