static/frontend: do not access elements until the page is loaded
When the js file is cached in the browser, it's very likely that
it will be executed before the DOM is ready. In this case, the
returned elements of a query is not the expected ones most of the
time. So the js code that accesses elements should be delayed
until the page is fully loaded.

Fixes golang/go#52502