tests/screentest: update snapshots
Updated the submenu test to include an inactive project
so that new releases of the standard library do
not cause the test to fail in the future. Additional
snapshots were update with a minor difference that
do not reach the threshold for a failed test.

Change-Id: I79d04bd3bb0ba36dd5df9c69952adf32355b2667
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/447740
Run-TryBot: Jamal Carvalho <jamal@golang.org>
TryBot-Result: kokoro <noreply+kokoro@google.com>
Auto-Submit: Jamal Carvalho <jamal@golang.org>
Reviewed-by: Hyang-Ah Hana Kim <hyangah@gmail.com>