internal/vuln: refactor vulndb client to prepare for v1 client
This change refactors the internal/vuln package to prepare for
the Client struct to support both the legacy
vulndb client and the new v1 client (introduced in a later CL).

The Client struct now contains only an embedded client interface which
is implemented by the legacyClient struct. The client interface is
essentially the same as the x/vuln Client interface, except that the
method GetByModule is replaced by ByPackage, which filters vulns
based on module, package and version.

The legacy client is refactored to implement ByPackage. In pkgsite,
ByPackage is used only by VulnsForPackage, whose behavior and interface
are not modified by this change. The rest of the legacy client is
essentially a pass-through to the x/vuln Client.

Once we have confidence that the v1 client works, the dependency on
the x/vuln Client will be removed completely.

Change-Id: Ia4f0d587ac4cfa28980f45ddbe8d609b5989666a
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/474695
Reviewed-by: Julie Qiu <julieqiu@google.com>
Reviewed-by: Tatiana Bradley <tatianabradley@google.com>
Run-TryBot: Tatiana Bradley <tatianabradley@google.com>
TryBot-Result: kokoro <noreply+kokoro@google.com>