devtools,tests/screentest: use vulndb v1 for screentests
Update screentests to run against the vulndb v1 schema instead of
the legacy schema by turning on the v1 experiment in the screentest
config.

Add a folder, vulndb-v1, containing test data for the v1 schema and
point to it for screentests. Also, add a script that can be used to
update this testdata.

Change-Id: Ia94ae4d125233e0c275b5abbfb22c2043e34f05f
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/479656
TryBot-Result: kokoro <noreply+kokoro@google.com>
Run-TryBot: Tatiana Bradley <tatianabradley@google.com>
Reviewed-by: Julie Qiu <julieqiu@google.com>