internal/frontend: add missing early return to retry fetch errors
A missing early return was causing retryable fetch errors not to be
retried by the frontend.

I spent a little while trying to test this, and it proved too difficult
and invasive. Unfortunately this may be a case where it is not
worthwhile to significantly refactor in order to improve testability.

Updates golang/go#59464

Change-Id: I0293a71d6be5046a4efe8b7ff490e285f912af3e
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/482162
Reviewed-by: Daniel Martí <mvdan@mvdan.cc>
Reviewed-by: Hyang-Ah Hana Kim <hyangah@gmail.com>
TryBot-Result: kokoro <noreply+kokoro@google.com>
Run-TryBot: Robert Findley <rfindley@google.com>