internal/worker: proxy timeouts should be retryable
When fetching e.g. the master branch, it is common for the proxy client
to timeout as it waits for the branch to be resolved.

This should be a retryable error, as otherwise we end up in a state
where an asynchronous refresh of master actually breaks documentation at
master.

There is still a problem that documentation is broken after the first
failure, but this at least puts us in a recoverable state.

Again, as with CL 482162 this is unfortunately not possible to test in
the current setup without significant refactoring. Such refactoring may
be warranted, but I do not have time to do this now.

Updates golang/go#59464

Change-Id: I321cc2eaabac1d7e052d07efcaadefcac24208ac
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/484736
Reviewed-by: Jamal Carvalho <jamal@golang.org>
Run-TryBot: Robert Findley <rfindley@google.com>
TryBot-Result: kokoro <noreply+kokoro@google.com>