internal/{frontend,vuln}: move logic to search by module to vuln client
Moves the logic for finding entries based on a module/package prefix
to the vuln client, so it is easier to test the core functionality.

Also modifies the functionality slightly, so that a search query
with a trailing "/" will show results.

Change-Id: Ief191a62bcf7badbc6e70112d9fcedae3dd03275
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/486456
Run-TryBot: Tatiana Bradley <tatianabradley@google.com>
Reviewed-by: Julie Qiu <julieqiu@google.com>
TryBot-Result: kokoro <noreply+kokoro@google.com>
Reviewed-by: Tatiana Bradley <tatianabradley@google.com>