cmd/frontend: add -local flag to run frontend in LocalMode
The frontend UI show/hide elements and links based on the fact that the frontend server is running in LocalMode or not. For the frontend this value is hardcoded to false. This commit add the flag -local to allow starting the frontend in localMode. This is particularly beneficial for those who wish to run the frontend locally or self-host for internal packages.