internal: fix build on Windows
Disable some of the vuln tests on windows because one of the path
conversion functions fails with an error saying it doesn't work on
Windows. Also fix the behavior on windows for /file paths. It looks
like they are intended to be /file/+filepath.ToSlash(filesystemPath)
when on Windows. Make sure we always generate those paths, and check
for those paths.

Change-Id: I23738b463b1415f2a75598dab4eaa511791f8e7e
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/508503
Run-TryBot: Michael Matloob <matloob@google.com>
TryBot-Result: kokoro <noreply+kokoro@google.com>
Reviewed-by: Jamal Carvalho <jamal@golang.org>